+++ noodles-vcf/src/io/writer/record/samples/sample/value/array.rs
@@ -113,9 +113,11 @@ mod tests {
         let array = ArrayBuf::String(vec![
             Some(String::from("noodles")),
             Some(String::from(";")),
+            Some(String::from("=")),
+            Some(String::from(":")),
             None,
         ]);
-        t(&mut buf, &array, b"noodles,%3B,.")?;
+        t(&mut buf, &array, b"noodles,;,=,%3A,.")?;
 
         Ok(())
     }