+++ noodles-vcf/CHANGELOG.md
@@ -7,8 +7,8 @@
   * vcf/io/writer/record/info/field/value: Remove colon (`:`) from encode set
     ([#339]).
 
-  * vcf/io/writer/record/samples/sample/value/character: Remove semicolon (`;`)
-    and equal sign (`=`) from encode set.
+  * vcf/io/writer/record/samples/sample/value: Remove semicolon (`;`) and equal
+    sign (`=`) from encode set.
 
 [#339]: https://github.com/zaeleus/noodles/issues/339